+++ src/model/index.js
@@ -1,3 +1,4 @@
 export { default as Loading } from './Loading';
+export { default as RatingTip } from './RatingTip';
 export { default as View } from './View';
 export { default as ShowType } from './ShowType';